user_service/requirements.txt
fastapi
uvicorn
databases
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose
requests
bcrypt==3.2.0
python-consul
pytest
grpcio==1.66.2
grpcio-tools==1.66.2
httpx
uuid
prometheus-fastapi-instrumentator